requirements.txt

certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
requests==2.28.1
slack-bolt==1.15.1
slack-sdk==3.19.1
urllib3==1.26.12
dbus-python==1.2.4
decorator==4.2.1
gpg==1.13.1
libcomps==0.1.18
nftables==0.1
perf==0.1
pygobject==3.28.3
python-dateutil==2.6.1
python-linux-procfs==0.7.0
pyudev==0.21.0
rpm==4.14.3
ruamel.yaml==0.17.21
selinux==2.9
six==1.11.0
slack-bolt==1.15.1
slip==0.6.4
slip.dbus==0.6.4
syspurpose==1.28.28
